specify type per command
Don't specify the type per parameter, but make it dependent on the
command and the parameter number.
This will save bytes and make implementing it a bit easier, especially
on the watch.

This commit also removes the double type, since it's not really needed.
Doubles can store integers up to 2^53 without any loss in precision.
Which is more than the 32 bit type we had earlier, of course it uses
the double amount of bits, but who cares?  Most of the readings are
going to be decimal numbers anyway.