use the watch name in the sidebar instead of the UID
The UID we get from the watch itself is _long_, it's probably an UUID of
some sort and those have to be long to be an UUID.

So let's just use the watch name we know and love, this change also sets
the watch name to `Watch <n>` were n is the number of the watch in the
list as of insertion.

Note: this can lead to duplicate names, but that isn't really a problem
in the code since we work with the IDs, but this can be confusing to
users.

In short: this change makes IDs more an implementation detail and makes
watch names the exposed thing for the user.